refactor(local-ci): extract windows validation script helper
Skill-Update: skip skill=ci reason="local-ci internal extraction only; no CI operator workflow change"
