+++ tools/local-ci/local_ci.py
@@ -3282,6 +3282,34 @@ def ps_literal(value: str) -> str:
     return _windows_probe.ps_literal(value)
 
 
+def windows_validation_script(
+    target_name: str,
+    host: str,
+    effective_repo_path: str,
+    job: dict,
+    *,
+    bundle_name: str,
+    bundle_ref: str,
+    exclude_tests: str,
+    cmake_generator: str,
+    resolved_platform: str,
+    resolved_generator_instance: str,
+) -> tuple[str, str]:
+    return _execution.windows_validation_script(
+        target_name,
+        host,
+        effective_repo_path,
+        job,
+        bundle_name=bundle_name,
+        bundle_ref=bundle_ref,
+        exclude_tests=exclude_tests,
+        cmake_generator=cmake_generator,
+        resolved_platform=resolved_platform,
+        resolved_generator_instance=resolved_generator_instance,
+        ps_literal_fn=ps_literal,
+    )
+
+
 def validate_ci_branch_name(branch: str) -> str:
     return _queue_orchestrator.validate_ci_branch_name(branch)
 
@@ -3467,352 +3495,18 @@ def run_windows_ssh_validation(
         cmake_generator_instance,
     )
 
-    ps_script = f"""
-$ErrorActionPreference = 'Stop'
-
-function Invoke-Native {{
-    param([string]$File, [string[]]$Arguments)
-    & $File @Arguments
-    if ($LASTEXITCODE -ne 0) {{
-        throw "$File exited with code $LASTEXITCODE"
-    }}
-}}
-
-function Test-CommitRef {{
-    param([string]$Ref)
-    & git rev-parse --verify --quiet "$Ref`^{{commit}}" 1> $null 2> $null
-    return $LASTEXITCODE -eq 0
-}}
-
-function Remove-DirectoryTreeRobust {{
-    param([string]$Path)
-
-    if (-not (Test-Path $Path)) {{
-        return
-    }}
-    try {{
-        cmd.exe /d /c ('rmdir /s /q "{{0}}"' -f $Path) | Out-Null
-    }} catch {{
-    }}
-    if (Test-Path $Path) {{
-        try {{
-            $LongPath = if ($Path.StartsWith('\\\\?\\')) {{ $Path }} else {{ '\\\\?\\' + $Path }}
-            Remove-Item -LiteralPath $LongPath -Recurse -Force -ErrorAction Stop
-        }} catch {{
-        }}
-    }}
-    if (Test-Path $Path) {{
-        try {{
-            Remove-Item -Recurse -Force -ErrorAction Stop $Path
-        }} catch {{
-        }}
-    }}
-}}
-
-function Remove-WorktreeSafe {{
-    param([string]$RepoRoot, [string]$Path)
-    try {{
-        Invoke-Native git @('-C', $RepoRoot, 'worktree', 'remove', '--force', '--force', $Path)
-    }} catch {{
-    }}
-    Remove-DirectoryTreeRobust $Path
-    try {{
-        Invoke-Native git @('-C', $RepoRoot, 'worktree', 'prune', '--expire', 'now')
-    }} catch {{
-    }}
-}}
-
-function Remove-PreparedRoot {{
-    param([string]$RepoRoot, [string]$PreparedRoot)
-
-    $PreparedSrc = Join-Path $PreparedRoot 'src'
-    if (Test-Path $PreparedSrc) {{
-        Remove-WorktreeSafe $RepoRoot $PreparedSrc
-    }}
-    if (Test-Path $PreparedRoot) {{
-        Remove-DirectoryTreeRobust $PreparedRoot
-    }}
-}}
-
-function Test-PreparedStateMatches {{
-    param(
-        [string]$StatePath,
-        [string]$ExpectedSha,
-        [string]$ExpectedValidation,
-        [string]$ExpectedGenerator,
-        [string]$ExpectedPlatform,
-        [string]$ExpectedGeneratorInstance
+    ps_script, validation = windows_validation_script(
+        target_name,
+        host,
+        effective_repo_path,
+        job,
+        bundle_name=bundle_name,
+        bundle_ref=bundle_ref,
+        exclude_tests=exclude_tests,
+        cmake_generator=cmake_generator,
+        resolved_platform=resolved_platform,
+        resolved_generator_instance=resolved_generator_instance,
     )
-
-    if (-not (Test-Path $StatePath)) {{
-        return $false
-    }}
-
-    try {{
-        $state = Get-Content $StatePath -Raw | ConvertFrom-Json
-    }} catch {{
-        return $false
-    }}
-
-    if (
-        $state.sha -ne $ExpectedSha -or
-        $state.validation -ne $ExpectedValidation -or
-        $state.generator -ne $ExpectedGenerator -or
-        $state.platform -ne $ExpectedPlatform -or
-        $state.generator_instance -ne $ExpectedGeneratorInstance
-    ) {{
-        return $false
-    }}
-
-    $PreparedRoot = Split-Path $StatePath -Parent
-    $PreparedSrc = Join-Path $PreparedRoot 'src'
-    $PreparedBuild = Join-Path $PreparedRoot 'build'
-    $PreparedInstall = Join-Path $PreparedRoot 'install'
-    if (-not (Test-Path $PreparedSrc) -or -not (Test-Path $PreparedBuild) -or -not (Test-Path $PreparedInstall)) {{
-        return $false
-    }}
-
-    $preparedHead = ((& git -C $PreparedSrc rev-parse HEAD 2>$null) | Select-Object -Last 1).Trim()
-    if ($LASTEXITCODE -ne 0) {{
-        return $false
-    }}
-    return $preparedHead -eq $ExpectedSha
-}}
-
-function Write-PreparedState {{
-    param(
-        [string]$StatePath,
-        [string]$Sha,
-        [string]$Validation,
-        [string]$Generator,
-        [string]$Platform,
-        [string]$GeneratorInstance
-    )
-
-    $payload = @{{
-        sha = $Sha
-        validation = $Validation
-        generator = $Generator
-        platform = $Platform
-        generator_instance = $GeneratorInstance
-        updated_at = (Get-Date).ToString('o')
-    }}
-    $payload | ConvertTo-Json | Set-Content -Path $StatePath
-}}
-
-function Wait-HostMutex {{
-    param(
-        [System.Threading.Mutex]$Mutex,
-        [bool]$Immediate
-    )
-
-    try {{
-        if ($Immediate) {{
-            return $Mutex.WaitOne(0)
-        }}
-        $null = $Mutex.WaitOne()
-        return $true
-    }} catch [System.Threading.AbandonedMutexException] {{
-        Write-Host "Recovered abandoned host validation lock: $MutexName"
-        return $true
-    }}
-}}
-
-$Repo = '{ps_literal(effective_repo_path)}'
-$RepoDrive = Split-Path -Path $Repo -Qualifier
-if (-not $RepoDrive) {{
-    $RepoDrive = 'C:'
-}}
-$env:GIT_LFS_SKIP_SMUDGE = '1'
-$CiRoot = Join-Path $RepoDrive 'pulp-ci'
-$Branch = '{ps_literal(job['branch'])}'
-$Sha = '{ps_literal(job['sha'])}'
-$BundleName = '{ps_literal(bundle_name)}'
-$BundleRef = '{ps_literal(bundle_ref)}'
-$Bundle = if ($BundleName) {{ Join-Path $HOME $BundleName }} else {{ '' }}
-$BundleGit = $Bundle.Replace('\\', '/')
-$ExcludeRegex = '{ps_literal(exclude_tests)}'
-$Generator = '{ps_literal(cmake_generator)}'
-$Platform = '{ps_literal(resolved_platform)}'
-$GeneratorInstance = '{ps_literal(resolved_generator_instance)}'
-$ValidationMode = '{ps_literal(job.get("validation", "full"))}'
-$PreparedRoot = Join-Path $CiRoot 'prepared\\{ps_literal(target_name)}'
-$PreparedRoot = Join-Path $PreparedRoot $ValidationMode
-$PreparedState = Join-Path $PreparedRoot 'state.json'
-$Src = Join-Path $PreparedRoot 'src'
-$Build = Join-Path $PreparedRoot 'build'
-$Install = Join-Path $PreparedRoot 'install'
-$Smoke = Join-Path $PreparedRoot 'smoke'
-$ReusePrepared = {'$true' if should_reuse_prepared_state(job) else '$false'}
-$UsePrepared = $false
-$MutexName = 'Global\\PulpLocalCIValidate'
-$Mutex = New-Object System.Threading.Mutex($false, $MutexName)
-$LockAcquired = $false
-$ValidatorStartedAt = (Get-Process -Id $PID).StartTime.ToUniversalTime().ToString('o')
-
-try {{
-    Write-Host "__PULP_VALIDATOR_PID__:$PID"
-    Write-Host "__PULP_VALIDATOR_STARTED__:$ValidatorStartedAt"
-    Write-Host "__PULP_VALIDATION__:$ValidationMode"
-    if ($ValidationMode -eq 'smoke') {{
-        Write-Host "__PULP_TEST_POLICY__:skip"
-    }} else {{
-        Write-Host "__PULP_TEST_POLICY__:run"
-    }}
-    if (-not (Wait-HostMutex -Mutex $Mutex -Immediate $true)) {{
-        Write-Host "__PULP_WAIT__:host-lock"
-        Write-Host "__PULP_PHASE__:waiting-lock"
-        Write-Host "Waiting for host validation lock: $MutexName"
-        $LockAcquired = Wait-HostMutex -Mutex $Mutex -Immediate $false
-    }} else {{
-        $LockAcquired = $true
-    }}
-
-    Write-Host "__PULP_PHASE__:fetch"
-    New-Item -ItemType Directory -Force -Path $PreparedRoot | Out-Null
-    Set-Location $Repo
-    if (Test-Path $Bundle) {{
-        Write-Host "__PULP_PHASE__:bundle-sync"
-        try {{
-            Invoke-Native git @(
-                'fetch',
-                $BundleGit,
-                "$BundleRef`:refs/pulp-ci-bundles/{job['id']}"
-            )
-        }} finally {{
-            Remove-Item -Force -ErrorAction SilentlyContinue $Bundle
-        }}
-    }}
-    if (-not (Test-CommitRef $Sha)) {{
-        try {{
-            Invoke-Native git @('fetch', 'origin')
-        }} catch {{
-        }}
-    }}
-
-    if (-not (Test-CommitRef $Sha)) {{
-        try {{
-            Invoke-Native git @(
-                'fetch',
-                'origin',
-                "refs/heads/$Branch`:refs/remotes/origin/$Branch"
-            )
-        }} catch {{
-        }}
-    }}
-
-    if (-not (Test-CommitRef $Sha)) {{
-        throw '{ps_literal(remote_commit_error(target_name, host, job))}'
-    }}
-
-    if ($ReusePrepared -and (Test-PreparedStateMatches `
-        -StatePath $PreparedState `
-        -ExpectedSha $Sha `
-        -ExpectedValidation $ValidationMode `
-        -ExpectedGenerator $Generator `
-        -ExpectedPlatform $Platform `
-        -ExpectedGeneratorInstance $GeneratorInstance)) {{
-        $UsePrepared = $true
-        Write-Host "__PULP_PREPARED__:reused"
-    }} else {{
-        Write-Host "__PULP_PREPARED__:clean"
-        Remove-PreparedRoot $Repo $PreparedRoot
-        New-Item -ItemType Directory -Force -Path $PreparedRoot | Out-Null
-        Write-Host "__PULP_PHASE__:worktree"
-        Invoke-Native git @('worktree', 'add', '--force', '--detach', $Src, $Sha)
-    }}
-
-    try {{
-        Write-Host "__PULP_PHASE__:configure"
-        Write-Host "CMake platform: $Platform"
-        if ($GeneratorInstance) {{
-            Write-Host "CMake generator instance: $GeneratorInstance"
-        }}
-        $configureArgs = @('-S', $Src, '-B', $Build)
-        if ($Generator) {{
-            $configureArgs += @('-G', $Generator)
-        }}
-        if ($Platform) {{
-            $configureArgs += @('-A', $Platform)
-        }}
-        if ($GeneratorInstance) {{
-            $configureArgs += @("-DCMAKE_GENERATOR_INSTANCE=$GeneratorInstance")
-        }}
-        $configureArgs += @('-DCMAKE_BUILD_TYPE=Release')
-        if ($ValidationMode -eq 'smoke') {{
-            $configureArgs += @(
-                '-DPULP_BUILD_TESTS=OFF',
-                '-DPULP_BUILD_EXAMPLES=OFF',
-                '-DPULP_ENABLE_GPU=OFF'
-            )
-        }}
-        Invoke-Native cmake $configureArgs
-        Write-Host "__PULP_PHASE__:build"
-        Invoke-Native cmake @('--build', $Build, '--config', 'Release')
-        if ($ValidationMode -eq 'smoke') {{
-            Write-Host "__PULP_PHASE__:install"
-            Invoke-Native cmake @('--install', $Build, '--prefix', $Install, '--config', 'Release')
-            New-Item -ItemType Directory -Force -Path $Smoke | Out-Null
-            @"
-cmake_minimum_required(VERSION 3.24)
-project(PulpSDKSmoke LANGUAGES CXX)
-
-find_package(Pulp REQUIRED CONFIG)
-
-add_library(smoke INTERFACE)
-target_link_libraries(smoke INTERFACE Pulp::format Pulp::standalone)
-"@ | Set-Content -Path (Join-Path $Smoke 'CMakeLists.txt')
-            Write-Host "__PULP_PHASE__:smoke"
-            $smokeConfigureArgs = @('-S', $Smoke, '-B', (Join-Path $Smoke 'build'))
-            if ($Generator) {{
-                $smokeConfigureArgs += @('-G', $Generator)
-            }}
-            if ($Platform) {{
-                $smokeConfigureArgs += @('-A', $Platform)
-            }}
-            if ($GeneratorInstance) {{
-                $smokeConfigureArgs += @("-DCMAKE_GENERATOR_INSTANCE=$GeneratorInstance")
-            }}
-            $smokeConfigureArgs += @("-DCMAKE_PREFIX_PATH=$Install")
-            Invoke-Native cmake $smokeConfigureArgs
-            Write-PreparedState `
-                -StatePath $PreparedState `
-                -Sha $Sha `
-                -Validation $ValidationMode `
-                -Generator $Generator `
-                -Platform $Platform `
-                -GeneratorInstance $GeneratorInstance
-        }} else {{
-            Write-PreparedState `
-                -StatePath $PreparedState `
-                -Sha $Sha `
-                -Validation $ValidationMode `
-                -Generator $Generator `
-                -Platform $Platform `
-                -GeneratorInstance $GeneratorInstance
-            Write-Host "__PULP_PHASE__:test"
-            $ctestArgs = @('--test-dir', $Build, '--output-on-failure', '-C', 'Release')
-            if ($ExcludeRegex) {{
-                $ctestArgs += @('--exclude-regex', $ExcludeRegex)
-            }}
-            Invoke-Native ctest $ctestArgs
-        }}
-    }} finally {{
-        Write-Host "__PULP_PHASE__:cleanup"
-        if (-not (Test-Path $PreparedState)) {{
-            Remove-PreparedRoot $Repo $PreparedRoot
-        }}
-    }}
-}} finally {{
-    if ($LockAcquired) {{
-        try {{
-            $Mutex.ReleaseMutex() | Out-Null
-        }} catch [System.ApplicationException] {{
-        }}
-    }}
-    $Mutex.Dispose()
-}}
-""".strip()
 
     cmd = windows_ssh_powershell_command(host)
 
@@ -3823,7 +3517,6 @@ target_link_libraries(smoke INTERFACE Pulp::format Pulp::standalone)
         log_path=log_path,
         report_progress=report_progress,
     )
-    validation = job.get("validation", "full")
     return validation_result_from_run(
         target_name,
         run,
